Tweak: Move overrides back to tailwind stylesheet
Moving overrides at the bottom of the tailwind stylesheet, after the
tailwind imports, removes the need for loading an additional stylesheet
to the theme.

These styles seem to be kept and added to the end of the compiled
stylesheet so this is fine for the time being.

This also removes the custom woocommerce.css enqueue and stylesheet.