Revert custom field name format
Reverted back to 'event_name' from 'event-name'. While I like the latter
a bit more, the former was how I originally had it set and the data was
already input in the DB under that name.

If I change, we will either have to re-add the date fields or I'll have
to go in and update the DB. It's easier to revert the name and it really
doesn't make a difference.