Add Custom Fields for Event and Webinar Dates (#20)
* Configure event date custom fields with Fieldmanager plugin.

* Add link to current site from COMING SOON page.

* Temp comment out date output to template.

* Output event date from custom field.

* Revert custom field name format

Reverted back to 'event_name' from 'event-name'. While I like the latter
a bit more, the former was how I originally had it set and the data was
already input in the DB under that name.

If I change, we will either have to re-add the date fields or I'll have
to go in and update the DB. It's easier to revert the name and it really
doesn't make a difference.

* Output upcoming & past events by event date.

* Set up query for featured event in the grid

This will allow for us to use a `featured` tag to display the featured
upcoming event in the top spot. It will only display the latest one.

Next we will write a separate query for the rest of the events and also
mix in the webinars.

* Update Featured Event query to order by date

This can act as a fallback to ensure only the latest featured event will
appear, based on the event date.

* Clean up unneeded class output logic.

* WIP: Set up custom query for home events

Still need to work in the Webinars. The Events are being pulled in fine
but the webinars are not showing up. Continuing to investigate the
query.

* Add Featured Event tag for featured item.