Adjust padding and positioning of nav items.
Increase the vertical padding for the main nav parent items. We will
also need to increase the horizontal padding but that can be done in a
follow-up.

Adjust the position of the dropdown lists to be below the other header
elements when visible and not cover them up.

Next steps:
1. Adjust horizontal padding of main parent nav items.
2. Add background color to the hovered parent and dropdown list.
3. Set nav lists to be horizontal when visible.